+++ promptolution/predictors/__init__.py
@@ -34,6 +34,6 @@ def get_predictor(name, *args, **kwargs):
     if name == "dummy":
         return DummyPredictor("", *args, **kwargs)
 
-    downstream_llm = get_llm(name)  # , batch_size=config.downstream_bs)
+    downstream_llm = get_llm(name)
 
     return Classificator(downstream_llm, *args, **kwargs)