+++ promptolution/tasks/__init__.py
@@ -7,22 +7,28 @@ from promptolution.tasks.base_task import BaseTask, DummyTask
 from promptolution.tasks.classification_tasks import ClassificationTask
 
 
-def get_tasks(config, split: Literal["dev", "test"] = "dev") -> List[BaseTask]:
-    """Create and return a list of task instances based on the provided configuration.
+def get_task(
+    config=None,
+    split: Literal["dev", "test"] = "dev",
+    ds_path: str = None,
+    task_name: str = None,
+    random_seed: int = None,
+) -> BaseTask:
+    """Create and return an task instance.
 
     This function supports creating multiple tasks, including a special 'dummy' task
-    for testing purposes and classification tasks based on JSON descriptions.
+    for testing purposes and classification tasks based on parsed config, or alternativly
+    the parsed arguments.
 
     Args:
-        config: Configuration object containing task settings.
-                Expected attributes:
-                - task_name (str): Comma-separated list of task names.
-                - ds_path (str): Path to the dataset directory.
-                - random_seed (int): Seed for random number generation.
-        split (Literal["dev", "test"], optional): Dataset split to use. Defaults to "dev".
+        config (Config): Configuration object containing the task details.
+        split (str): Split of the dataset to use for the task (default: 'dev').
+        ds_path (str): Path to the dataset containing the task description.
+        task_name (str): Name of the task to create.
+        random_seed (int): Random seed for the task.
 
     Returns:
-        List[BaseTask]: A list of instantiated task objects.
+        BaseTask: A list of instantiated task objects.
 
     Raises:
         FileNotFoundError: If the task description file is not found.
@@ -33,17 +39,17 @@ def get_tasks(config, split: Literal["dev", "test"] = "dev") -> List[BaseTask]:
         - For all other tasks, a ClassificationTask instance is created.
         - The task description is loaded from a 'description.json' file in the dataset path.
     """
-    task_names = config.task_name.split(",")
+    if config.task_name == "dummy":
+        task = DummyTask()
+        return task
 
-    task_list = []
-    for task_name in task_names:
-        task_description_path = Path(config.ds_path) / Path("description.json")
-        task_description = json.loads(task_description_path.read_text())
-        if task_name == "dummy":
-            task = DummyTask()
-            task_list.append(task)
-            continue
-        task = ClassificationTask(task_name, task_description, split=split, seed=config.random_seed)
-        task_list.append(task)
+    if ds_path is None:
+        ds_path = config.ds_path
+    if task_name is None:
+        task_name = config.task_name
+    if random_seed is None:
+        random_seed = config.random_seed
+    task_description_path = Path(ds_path)
+    task = ClassificationTask(task_description_path, task_name, split=split, seed=random_seed)
 
-    return task_list
+    return task