+++ promptolution/tasks/classification_tasks.py
@@ -1,9 +1,11 @@
 """Module for classification tasks."""
 
+import json
 from pathlib import Path
-from typing import Dict, List, Literal, Optional
+from typing import Callable, Dict, List, Literal, Optional
 
 import numpy as np
+from sklearn.metrics import accuracy_score
 
 from promptolution.predictors.base_predictor import BasePredictor
 from promptolution.tasks.base_task import BaseTask
@@ -17,36 +19,48 @@ class ClassificationTask(BaseTask):
 
     Attributes:
         task_id (str): Unique identifier for the task.
+        path (Path): Path to the dataset description JSON file, and initial prompts.
         dataset_json (Dict): Dictionary containing dataset information.
         description (Optional[str]): Description of the task.
         initial_population (Optional[List[str]]): Initial set of prompts.
         xs (Optional[np.ndarray]): Input data for the task.
         ys (Optional[np.ndarray]): Ground truth labels for the task.
         classes (Optional[List]): List of possible class labels.
-        split (Literal["dev", "test"]): Dataset split to use.
         seed (int): Random seed for reproducibility.
+        split (Literal["dev", "test"]): Dataset split to use.
+        metric (Callable): Metric to use as an evaluation score for the prompts.
 
     Inherits from:
         BaseTask: The base class for tasks in the promptolution library.
     """
 
-    def __init__(self, task_id: str, dataset_json: Dict, seed: int = 42, split: Literal["dev", "test"] = "dev"):
+    def __init__(
+        self,
+        dataset_path: Path,
+        task_id: str = "Classification Task",
+        seed: int = 42,
+        split: Literal["dev", "test"] = "dev",
+        metric: Callable = accuracy_score,
+    ):
         """Initialize the ClassificationTask.
 
         Args:
             task_id (str): Unique identifier for the task.
-            dataset_json (Dict): Dictionary containing dataset information.
+            dataset_path (str): Path to the dataset description JSON file.
             seed (int, optional): Random seed for reproducibility. Defaults to 42.
             split (Literal["dev", "test"], optional): Dataset split to use. Defaults to "dev".
+            metric (Callable): Metric to use as an evaluation score for the prompts. Defaults to sklearn's accuracy.
         """
         self.task_id: str = task_id
-        self.dataset_json: Dict = dataset_json
+        self.path: Path = dataset_path
+        self.dataset_json: Dict = json.loads((dataset_path / Path("description.json")).read_text())
         self.description: Optional[str] = None
         self.initial_population: Optional[List[str]] = None
         self.xs: Optional[np.ndarray] = np.array([])
         self.ys: Optional[np.ndarray] = None
         self.classes: Optional[List] = None
         self.split: Literal["dev", "test"] = split
+        self.metric = metric
         self._parse_task()
         self.reset_seed(seed)
 
@@ -60,18 +74,17 @@ class ClassificationTask(BaseTask):
         This method loads the task description, classes, initial prompts,
         and the dataset split (dev or test) into the class attributes.
         """
-        task_path = Path(self.dataset_json["path"])
         self.description = self.dataset_json["description"]
         self.classes = self.dataset_json["classes"]
 
-        with open(task_path / Path(self.dataset_json["init_prompts"]), "r", encoding="utf-8") as file:
+        with open(self.path / Path(self.dataset_json["init_prompts"]), "r", encoding="utf-8") as file:
             lines = file.readlines()
         self.initial_population = [line.strip() for line in lines]
 
         seed = Path(self.dataset_json["seed"])
         split = Path(self.split + ".txt")
 
-        with open(task_path / seed / split, "r", encoding="utf-8") as file:
+        with open(self.path / seed / split, "r", encoding="utf-8") as file:
             lines = file.readlines()
         lines = [line.strip() for line in lines]
 
@@ -87,7 +100,12 @@ class ClassificationTask(BaseTask):
         self.ys = np.array(ys)
 
     def evaluate(
-        self, prompts: List[str], predictor: BasePredictor, n_samples: int = 20, subsample: bool = True
+        self,
+        prompts: List[str],
+        predictor: BasePredictor,
+        n_samples: int = 20,
+        subsample: bool = False,
+        return_seq: bool = False,
     ) -> np.ndarray:
         """Evaluate a set of prompts using a given predictor.
 
@@ -95,7 +113,9 @@ class ClassificationTask(BaseTask):
             prompts (List[str]): List of prompts to evaluate.
             predictor (BasePredictor): Predictor to use for evaluation.
             n_samples (int, optional): Number of samples to use if subsampling. Defaults to 20.
-            subsample (bool, optional): Whether to use subsampling. Defaults to True.
+            subsample (bool, optional): Whether to use subsampling.
+            If set to true, samples a different subset per call. Defaults to False.
+            return_seq (bool, optional): whether to return the generating sequence
 
         Returns:
             np.ndarray: Array of accuracy scores for each prompt.
@@ -112,10 +132,17 @@ class ClassificationTask(BaseTask):
         ys_subsample = self.ys[indices]
 
         # Make predictions on the subsample
-        preds = predictor.predict(prompts, xs_subsample)
+        preds = predictor.predict(prompts, xs_subsample, return_seq=return_seq)
 
-        # Calculate accuracy: number of correct predictions / total number of predictions per prompt
-        return np.mean(preds == ys_subsample, axis=1)
+        if return_seq:
+            preds, seqs = preds
+
+        scores = np.array([self.metric(ys_subsample, pred) for pred in preds])
+
+        if return_seq:
+            return scores, seqs
+
+        return scores
 
     def reset_seed(self, seed: int = None):
         """Reset the random seed."""